+++ src/assets/styles/bootstrap.css
@@ -330,10 +330,9 @@ img {
 }
 
 hr {
-  margin-top: 20px;
   margin-bottom: 20px;
   border: 0;
-  border-top: 1px solid #eeeeee;
+  border-top: 2px solid #eeeeee;
 }
 
 .sr-only {
