+++ src/components/index.jsx
@@ -1,0 +1,1 @@
+export * from './PageHeader';
