+++ src/pages/Info/Empresa.jsx
@@ -1,0 +1,9 @@
+export const Empresa = () => (
+  <>
+    <div>
+      A Printástico é uma empresa brasileira especializada em FotoProdutos e impressões digitais, que segue os mais altos padrões internacionais de tecnologia e qualidade. Nossa missão é oferecer aos nossos clientes um ambiente seguro e aplicações que garantam a criação de projetos personalizados de maneira rápida, fácil e divertida.</div><div><br /></div><div>Nosso alto padrão é possível devido à expertise de nossa equipe, formada por consultores de diversos países como EUA, Alemanha, Suécia, Inglaterra, Itália, Dinamarca, Israel, Eslováquia e outros. Além disso, contamos com o suporte de nossa equipe técnica qualificada que garante a excelência de nossos serviços.</div><div><br /></div><div>Por sermos uma empresa online, todas as nossas aplicações podem ser acessadas a qualquer momento do dia. Os pedidos também seguem a mesma lógica, ou seja, também são processados 24 horas por dia.</div><div><br /></div><div>Trabalhamos de forma a transformar suas melhores memórias em projetos bonitos, com alto padrão de qualidade. Tudo para que sua história possa ser contada e recontada quantas vezes você, sua família e seus amigos quiserem.
+    </div>
+    <br />
+    <div>O nosso compromisso inclui ainda oferecer sempre novidades que alinhem beleza, praticidade e bons preços.</div>
+  </>
+);
