+++ src/pages/Info/index.jsx
@@ -1,1 +1,2 @@
+export * from './Info';
 export * from './Empresa';
