+++ src/pages/index.jsx
@@ -1,3 +1,2 @@
 export * from './Lp';
-export * from './Page1';
-export * from './Page2';
+export * from './Info';
