--- src/pages/informations/Empresa.jsx
@@ -1,45 +1,0 @@
-import './index.css';
-
-export const Empresa = () => (
-  <>
-    {/* <!-- BREADCRUMB ================================================== --> */}
-    <div className="breadcrumb full-width">
-      <div className="background-breadcrumb"></div>
-      <div className="background">
-        <div className="shadow"></div>
-        <div className="pattern">
-          <div className="container">
-            <div className="clearfix" style={{ paddingTop: '50px' }}>
-              <h1 id="title-page">A Empresa					     						     </h1>
-              <ul>
-                <li><a href="/">Home</a></li>
-                <li><a href="http://opencart2.printastico.com.br/a-empresa">A Empresa</a></li>
-              </ul>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    {/* <!-- MAIN CONTENT ================================================== --> */}
-    <div className="main-content full-width inner-page">
-      <div className="background-content"></div>
-      <div className="background">
-        <div className="shadow"></div>
-        <div className="pattern">
-          <div className="container">
-            <div style={{ display: 'flex', justifyContent: 'center' }}>
-              <div className="col-md-9 center-column content-with-background" id="content" style={{ marginTop: '10px' }}>
-                <div style={{ paddingBottom: '10px', fontSize: '13px' }}><div>A Printástico é uma empresa brasileira especializada em FotoProdutos e impressões digitais, que segue os mais altos padrões internacionais de tecnologia e qualidade. Nossa missão é oferecer aos nossos clientes um ambiente seguro e aplicações que garantam a criação de projetos personalizados de maneira rápida, fácil e divertida.</div><div><br /></div><div>Nosso alto padrão é possível devido à expertise de nossa equipe, formada por consultores de diversos países como EUA, Alemanha, Suécia, Inglaterra, Itália, Dinamarca, Israel, Eslováquia e outros. Além disso, contamos com o suporte de nossa equipe técnica qualificada que garante a excelência de nossos serviços.</div><div><br /></div><div>Por sermos uma empresa online, todas as nossas aplicações podem ser acessadas a qualquer momento do dia. Os pedidos também seguem a mesma lógica, ou seja, também são processados 24 horas por dia.</div><div><br /></div><div>Trabalhamos de forma a transformar suas melhores memórias em projetos bonitos, com alto padrão de qualidade. Tudo para que sua história possa ser contada e recontada quantas vezes você, sua família e seus amigos quiserem.</div><div><br /></div><div>O nosso compromisso inclui ainda oferecer sempre novidades que alinhem beleza, praticidade e bons preços.</div></div>
-              </div>
-            </div>
-            <div className="row">
-              <div className="col-sm-12">
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </>
-);
