--- src/pages/page1.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const Page1Wrapper = styled.div`
-  background: green;
-  height: 500px;
-  width: 500px;
-`;
-
-export const Page1 = () => (
-  <Page1Wrapper>PAG 1</Page1Wrapper>
-);
