--- src/pages/page2.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const Page1Wrapper = styled.div`
-  background: yellow;
-  height: 700px;
-  width: 500px;
-`;
-
-export const Page2 = () => (
-  <Page1Wrapper>PAG 2</Page1Wrapper>
-);
