+++ src/routes/routes.jsx
@@ -1,17 +1,12 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import {
-  Lp,
-  Page1,
-  Page2,
-} from '@/pages';
-import { Empresa } from '@/pages/informations';
+import { Lp, Info, Empresa } from '@/pages';
 
 export function AppRoutes() {
   return (
     <Routes>
       <Route path="/" element={<Lp />} />
-      <Route path="/a-empresa" element={<Empresa />} />
+      <Route path="/info/:infoType" element={<Info />} />
     </Routes>
   );
 }
